Logo component rewritten to TS
