+++ src/components/Logo/Logo.test.tsx
@@ -1,6 +1,5 @@
 import { LongLogo, ShortLogo } from '@components/Logo'
 import { render } from '@testing-library/react'
-import React from 'react'
 
 describe('<ShortLogo/> mount', () => {
   it('is mounted', () => {
