+++ src/components/Logo/index.ts
